Remove slow MoE path, make fast path mandatory
- Delete _forward_slow() method (~90 lines)
- Replace fallback in forward_fast() with RuntimeError
- Simplify forward() to always call forward_fast()
- Remove _use_fast_moe, backoff fields and failure tracking
- Add mixed-precision detection (_is_mixed_precision)
- Add _forward_grouped() for per-bit-group dispatch
- Update tests to remove _use_fast_moe checks
- Update docstrings: forward_fast -> forward

Breaking change: Models must be properly initialized with
_create_buffers_eagerly() before inference. CPU tensors no longer
supported - all MoE ops require MPS device.

This simplifies the codebase and eliminates confusion between
fast/slow paths. The 'slow' path was too slow (~0.5 tok/s) to be
useful as a fallback.

+++ metal_marlin/trellis/dispatch.py
@@ -56,7 +56,7 @@ def dispatch_trellis_dequant(
 
     Args:
         lib: MetalKernelLibrary with dequant_trellis compiled
-        indices: Trellis indices [tiles_k, tiles_n, 256] int16, MPS tensor
+        indices: Trellis indices [tiles_k, tiles_n, 256] uint8, MPS tensor
         scales: Per-group scales [n_groups, N] float32, MPS tensor
         grid: Codebook grid [n_levels] float32, MPS tensor
         K: Number of rows in output
@@ -193,7 +193,7 @@ def dispatch_trellis_dequant_fused(
 
     Args:
         lib: MetalKernelLibrary with dequant_trellis compiled
-        indices: Trellis indices [tiles_k, tiles_n, 256] int16, MPS tensor
+        indices: Trellis indices [tiles_k, tiles_n, 256] uint8, MPS tensor
         scales: Per-group scales [n_groups, N] float32, MPS tensor
         grid: Codebook grid [n_levels] float32, MPS tensor
         su: Row sign vector [K] float32, MPS tensor
